requirements.txt
pandas
python-dotenv
gtts
moviepy
openai
google-api-python-client
